+++ db-queue-scheduler-core/src/test/java/ru/yoomoney/tech/dbqueue/scheduler/internal/queue/ScheduledTaskQueueConsumerTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.Test;
 import ru.yoomoney.tech.dbqueue.api.Task;
 import ru.yoomoney.tech.dbqueue.api.TaskExecutionResult;
 import ru.yoomoney.tech.dbqueue.config.QueueShardId;
+import ru.yoomoney.tech.dbqueue.scheduler.config.ScheduledTaskLifecycleListener;
+import ru.yoomoney.tech.dbqueue.scheduler.config.impl.NoopScheduledTaskLifecycleListener;
 import ru.yoomoney.tech.dbqueue.scheduler.internal.ScheduledTaskDefinition;
 import ru.yoomoney.tech.dbqueue.scheduler.internal.schedule.impl.FixedRateNextExecutionTimeProvider;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskExecutionResult;
@@ -21,6 +23,8 @@ import ru.yoomoney.tech.dbqueue.settings.QueueSettings;
 import ru.yoomoney.tech.dbqueue.settings.ReenqueueRetryType;
 import ru.yoomoney.tech.dbqueue.settings.ReenqueueSettings;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
@@ -51,7 +55,8 @@ class ScheduledTaskQueueConsumerTest {
                 .build();
         ScheduledTaskQueueConsumer scheduledTaskQueueConsumer = new ScheduledTaskQueueConsumer(
                 dummyQueueConfig(),
-                scheduledTaskDefinition
+                scheduledTaskDefinition,
+                NoopScheduledTaskLifecycleListener.getInstance()
         );
 
         // when
@@ -74,7 +79,8 @@ class ScheduledTaskQueueConsumerTest {
                 .build();
         ScheduledTaskQueueConsumer scheduledTaskQueueConsumer = new ScheduledTaskQueueConsumer(
                 dummyQueueConfig(),
-                scheduledTaskDefinition
+                scheduledTaskDefinition,
+                NoopScheduledTaskLifecycleListener.getInstance()
         );
 
         // when
@@ -97,6 +103,7 @@ class ScheduledTaskQueueConsumerTest {
         ScheduledTaskQueueConsumer scheduledTaskQueueConsumer = new ScheduledTaskQueueConsumer(
                 dummyQueueConfig(),
                 scheduledTaskDefinition,
+                NoopScheduledTaskLifecycleListener.getInstance(),
                 clock
         );
 
@@ -109,7 +116,7 @@ class ScheduledTaskQueueConsumerTest {
     }
 
     @Test
-    public void should_postpone_execution_according_to_overrid() {
+    public void should_postpone_execution_according_to_overridden_next_execution_time() {
         // given
         Clock clock = Clock.fixed(Instant.now(), ZoneId.systemDefault());
         ScheduledTaskDefinition scheduledTaskDefinition = ScheduledTaskDefinition.builder()
@@ -122,6 +129,7 @@ class ScheduledTaskQueueConsumerTest {
         ScheduledTaskQueueConsumer scheduledTaskQueueConsumer = new ScheduledTaskQueueConsumer(
                 dummyQueueConfig(),
                 scheduledTaskDefinition,
+                NoopScheduledTaskLifecycleListener.getInstance(),
                 clock
         );
 
@@ -131,6 +139,41 @@ class ScheduledTaskQueueConsumerTest {
         // then
         assertThat(taskExecutionResult.getActionType(), equalTo(TaskExecutionResult.Type.REENQUEUE));
         assertThat(taskExecutionResult.getExecutionDelay().get(), equalTo(Duration.ofDays(10L)));
+    }
+
+    @Test
+    @SuppressWarnings("AccessingNonPublicFieldOfAnotherObject")
+    public void should_receive_lifecycle_events() {
+        // given
+        Clock clock = Clock.fixed(Instant.now(), ZoneId.systemDefault());
+        RuntimeException exception = new RuntimeException("test exception");
+        ScheduledTaskDefinition scheduledTaskDefinition = ScheduledTaskDefinition.builder()
+                .withScheduledTask(() -> {
+                    throw exception;
+                })
+                .withScheduledTaskIdentity(ScheduledTaskIdentity.of("scheduled-task"))
+                .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                .withNextExecutionTimeProvider(new FixedRateNextExecutionTimeProvider(Duration.ofDays(1L), clock))
+                .build();
+        DummyScheduledTaskLifecycleListener listener = new DummyScheduledTaskLifecycleListener();
+        ScheduledTaskQueueConsumer scheduledTaskQueueConsumer = new ScheduledTaskQueueConsumer(
+                dummyQueueConfig(),
+                scheduledTaskDefinition,
+                listener,
+                clock
+        );
+
+        // when
+        scheduledTaskQueueConsumer.execute(dummyTask());
+
+        // then
+        assertThat(listener.startedTaskIdentity, equalTo(ScheduledTaskIdentity.of("scheduled-task")));
+        assertThat(listener.finishedTaskIdentity, equalTo(ScheduledTaskIdentity.of("scheduled-task")));
+        assertThat(listener.crashedTaskIdentity, equalTo(ScheduledTaskIdentity.of("scheduled-task")));
+        assertThat(listener.executionResult, equalTo(ScheduledTaskExecutionResult.error()));
+        assertThat(listener.processTaskTimeInMills, equalTo(0L));
+        assertThat(listener.nextExecutionTime, equalTo(clock.instant().plus(Duration.ofDays(1L))));
+        assertThat(listener.throwable, equalTo(exception));
     }
 
     private Task<String> dummyTask() {
@@ -169,5 +212,37 @@ class ScheduledTaskQueueConsumerTest {
                         .withExtSettings(ExtSettings.builder().withSettings(Collections.emptyMap()).build())
                         .build()
         );
+    }
+
+    private static class DummyScheduledTaskLifecycleListener implements ScheduledTaskLifecycleListener {
+        private ScheduledTaskIdentity startedTaskIdentity;
+        private ScheduledTaskIdentity finishedTaskIdentity;
+        private ScheduledTaskIdentity crashedTaskIdentity;
+        private ScheduledTaskExecutionResult executionResult;
+        private Instant nextExecutionTime;
+        private Long processTaskTimeInMills;
+        private Throwable throwable;
+
+        @Override
+        public void started(@Nonnull ScheduledTaskIdentity taskIdentity) {
+            this.startedTaskIdentity = taskIdentity;
+        }
+
+        @Override
+        public void finished(@Nonnull ScheduledTaskIdentity taskIdentity,
+                             @Nonnull ScheduledTaskExecutionResult executionResult,
+                             @Nonnull Instant nextExecutionTime,
+                             long processTaskTimeInMills) {
+            this.finishedTaskIdentity = taskIdentity;
+            this.executionResult = executionResult;
+            this.nextExecutionTime = nextExecutionTime;
+            this.processTaskTimeInMills = processTaskTimeInMills;
+        }
+
+        @Override
+        public void crashed(@Nonnull ScheduledTaskIdentity taskIdentity, @Nullable Throwable exc) {
+            this.crashedTaskIdentity = taskIdentity;
+            this.throwable = exc;
+        }
     }
 }